@@ -1,28 +1,37 @@
 {
-  "name": "othello-ai",
+  "name": "othello-ai-with-llm",
   "private": true,
+  "version": "0.1.0",
+  "type": "module",
   "scripts": {
-    "build": "webpack --config webpack.prod.js",
-    "dev": "webpack-dev-server -d --progress --colors --config webpack.dev.js"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0 --fix",
+    "test": "vitest"
   },
   "devDependencies": {
-    "@types/lodash": "^4.14.85",
-    "@types/node": "^8.0.53",
-    "@types/react": "^16.0.25",
-    "@types/react-dom": "^16.0.3",
-    "@types/react-redux": "^5.0.14",
-    "@types/redux": "^3.6.0",
-    "@types/redux-logger": "^3.0.5",
-    "lodash": "^4.17.4",
-    "react": "^16.1.1",
-    "react-dom": "^16.1.1",
-    "react-redux": "^5.0.6",
-    "redux": "^3.7.2",
+    "@reduxjs/toolkit": "^2.2.2",
+    "@types/lodash": "^4.17.0",
+    "@types/node": "^20.11.30",
+    "@types/react": "^18.2.73",
+    "@types/react-dom": "^18.2.22",
+    "@types/redux-logger": "^3.0.13",
+    "@typescript-eslint/eslint-plugin": "^7.4.0",
+    "@typescript-eslint/parser": "^7.4.0",
+    "@vitejs/plugin-react": "^4.2.1",
+    "@vitejs/plugin-react-swc": "^3.6.0",
+    "eslint": "^8.57.0",
+    "eslint-plugin-react": "^7.34.1",
     "redux-logger": "^3.0.6",
-    "ts-loader": "^3.1.1",
-    "typescript": "^2.6.2",
-    "webpack": "^3.8.1",
-    "webpack-dev-server": "^2.9.5",
-    "webpack-merge": "^4.1.1"
+    "vite": "^5.2.6",
+    "vite-tsconfig-paths": "^4.3.2",
+    "vitest": "^1.4.0"
+  },
+  "dependencies": {
+    "lodash": "^4.17.21",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^9.1.0",
+    "redux": "^5.0.1"
   }
 }
