@@ -1,42 +1,16 @@
-import * as React from "react"
-import * as ReactDOM from "react-dom"
-import { Store, Middleware } from "redux"
-import { applyMiddleware, createStore } from "redux"
-import { createLogger } from "redux-logger"
-import { Provider } from "react-redux"
-import { Game, GameState } from "ui/containers/Game"
-import { reducers } from "ui/reducers"
-import * as Constants from "ui/constants"
-import { bench } from "bitboard/bench"
-
-const middleWares: Middleware[] = [
-    process.env.NODE_ENV !== "production" && createLogger() as any
-].filter(Boolean)
-
-export function start(dom: HTMLElement | null) {
-    let store = createStore<GameState>(
-        reducers,
-        {
-            positions: [{
-                cells: Constants.initialBoard,
-                turn: "b"
-            }],
-            playerColor: "b"
-        },
-        applyMiddleware(...middleWares)
-    )
-
-    ReactDOM.render(
-        <App store={store} />,
-        dom
-    )
-}
-
+import { Game, GameState } from "@app/ui/containers/Game";
+import { Provider } from "react-redux";
+import { Store } from "redux";
+import { Action } from "./actions";
 export interface AppProps {
-    store: Store<GameState>
+  store: Store<GameState, Action>;
 }
 const App = ({ store }: AppProps) => (
+  <>
     <Provider store={store}>
-        <Game />
+      <Game />
     </Provider>
-)
+  </>
+);
+
+export default App;
