@@ -1,8 +1,8 @@
-export const cellWidth = 40
-export const cellMargin = 1
-export const boardWidth = (cellWidth + cellMargin * 2) * 8
-export const boardMargin = 10
-export const controlHeight = 30
-export const mainWidth = boardWidth + boardMargin * 2
-export const mainHeight = boardWidth + boardMargin + controlHeight
-export const stoneMargin = 2
+export const cellWidth = 40;
+export const cellMargin = 1;
+export const boardWidth = (cellWidth + cellMargin * 2) * 8;
+export const boardMargin = 10;
+export const controlHeight = 30;
+export const mainWidth = boardWidth + boardMargin * 2;
+export const mainHeight = boardWidth + boardMargin + controlHeight;
+export const stoneMargin = 2;
