@@ -1,57 +1,65 @@
-import * as _ from "lodash"
-import { connect } from "react-redux"
-import { Dispatch } from "redux"
-import * as actions from "ui/actions"
-import { Main, MainProps } from "ui/components/Main"
-import { Color, CellState, Place } from "ui/types"
-import * as Move from "bitboard/move"
-import * as Board from "bitboard/Board"
+import * as Board from "@app/bitboard/Board";
+import * as Move from "@app/bitboard/move";
+import * as actions from "@app/ui/actions";
+import Main, { Status } from "@app/ui/components/Main";
+import { CellState, Color, Place } from "@app/ui/types";
+import * as _ from "lodash";
+import { connect } from "react-redux";
+import { Dispatch } from "redux";
 
 export interface GameState {
-    positions: Position[],
-    playerColor: Color,
-    latestMove?: Place
+  positions: Position[];
+  playerColor: Color;
+  latestMove?: Place;
 }
 
 export interface Position {
-    turn: Color,
-    cells: CellState[]
+  turn: Color;
+  cells: CellState[];
 }
 
-function mapStateToProps(state: GameState, ownProps: any) {
-    const position = _.last(state.positions) as Position
-    const board = position.turn == "b" ?
-        Board.fromUiState(position.cells) :
-        Board.reverse(Board.fromUiState(position.cells))
-    const [black, white] = position.turn == "b" ?
-        Board.stones(board) :
-        Board.stones(board).reverse()
-    let status = "normal"
-    if (Move.movables(board).length == 0) {
-        if (Move.movables(Board.reverse(board)).length == 0) {
-            status = "finished"
-        } else {
-            status = "pass"
-        }
+function mapStateToProps(state: GameState) {
+  const position = _.last(state.positions) as Position;
+  const board =
+    position.turn == "b"
+      ? Board.fromUiState(position.cells)
+      : Board.reverse(Board.fromUiState(position.cells));
+  const [black, white] =
+    position.turn == "b" ? Board.stones(board) : Board.stones(board).reverse();
+  let status: Status = "normal";
+  if (Move.movables(board).length == 0) {
+    if (Move.movables(Board.reverse(board)).length == 0) {
+      status = "finished";
+    } else {
+      status = "pass";
     }
-    return {
-        cells: position.cells,
-        turn: position.turn,
-        status,
-        black,
-        white,
-        playerColor: state.playerColor,
-        latestMove: state.latestMove
-    }
+  }
+  return {
+    cells: position.cells,
+    turn: position.turn,
+    status,
+    black,
+    white,
+    playerColor: state.playerColor,
+    latestMove: state.latestMove,
+  };
 }
 
-function mapDispatchToProps(dispatch: Dispatch<0>): {} {
-    return {
-        onClickCell(place: Place) { dispatch(actions.clickCell(place)) },
-        onClickPrev() { dispatch(actions.clickPrev) },
-        onClickPass() { dispatch(actions.clickPass) },
-        launchAi() { dispatch(actions.aiMove) }
-    }
+function mapDispatchToProps(dispatch: Dispatch<actions.BoardAction>) {
+  return {
+    onClickCell(place: Place) {
+      dispatch(actions.clickCell(place));
+    },
+    onClickPrev() {
+      dispatch(actions.clickPrev);
+    },
+    onClickPass() {
+      dispatch(actions.clickPass);
+    },
+    launchAi() {
+      dispatch(actions.aiMove);
+    },
+  };
 }
 
-export const Game: any = connect<any>(mapStateToProps, mapDispatchToProps)(Main)
+export const Game = connect(mapStateToProps, mapDispatchToProps)(Main);
