@@ -1,7 +1,7 @@
 export interface Place {
-    x: number,
-    y: number
+  x: number;
+  y: number;
 }
 
-export type CellState = "." | "b" | "w"
-export type Color = "b" | "w"
+export type CellState = "." | "b" | "w";
+export type Color = "b" | "w";
